fix: Wrap `SupportedDevicesTable` in `QueryClientProvider`
